requirements.txt
awscli==1.6.2
backports-abc==0.5
backports.ssl-match-hostname==3.5.0.1
bcdoc==0.12.2
boto==2.33.0
botocore==0.73.0
certifi==2018.4.16
chardet==3.0.4
colorama==0.2.5
decorator==4.0.9
diff-match-patch==20121119
Django==1.9.8
django-crispy-forms==1.7.2
django-formtools==2.1
django-import-export==1.0.0
django-pure-pagination==0.3.0
django-simple-captcha==0.4.6
DjangoUeditor==1.8.143
docutils==0.14
et-xmlfile==1.0.1
functools32==3.2.3.post2
future==0.16.0
genson==0.1.0
httplib2==0.9.2
idna==2.6
image==1.5.24
jdcal==1.4
jmespath==0.5.0
jsonschema==2.5.1
multipledispatch==0.5.0
MySQL-python==1.2.5
odfpy==1.3.6
openpyxl==2.5.3
Pillow==5.1.0
pyasn1==0.4.2
python-dateutil==2.7.3
PyYAML==3.12
requests==2.18.4
rsa==3.1.2
six==1.11.0
tablib==0.12.1
unicodecsv==0.14.1
urllib3==1.22
uWSGI==2.0.17
xlrd==1.1.0
xlwt==1.3.0
